+++ python/mujoco/usd/demo.py
@@ -20,7 +20,6 @@ import pathlib
 import mujoco
 from mujoco.usd import exporter
 
-
 def generate_usd_trajectory(local_args):
   """Generates a USD file given the user arguments."""
   # load a model to mujoco
@@ -43,6 +42,10 @@ def generate_usd_trajectory(local_args):
     mujoco.mj_step(m, d)
     if exp.frame_count < d.time * local_args.framerate:
       exp.update_scene(data=d, camera=cam)
+
+  exp.add_light(pos=(0, 0, 0),
+                intensity=2000,
+                light_type="dome")
 
   exp.save_scene(filetype=local_args.export_extension)